requirements-dev.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-allauth==0.43.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
idna==2.10
importlib-metadata==2.0.0
Markdown==3.3.1
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.25.10
zipp==3.3.0
django-cors-headers
